Initial Commit
Initial Commit of this projects late-mvp
Messages can help give directions on how to play.  Click on tiles to
select them, press space to spawn a unit near king.
Note: I intended this to be just a quick test on functionality but it
quickly became the mvp.
MVP - Minimal Viable Product